Add color to extract script, add extract_all

+++ tools/extract.php
@@ -2,7 +2,7 @@
 include_once 'lib/index.php';
 
 if ($argc < 2) {
-    echo "Usage: php extract.php <filename>\n";
+    echo "Usage: php extract.php <filename>" . PHP_EOL;
     exit(1);
 }
 $filename = $argv[1];
@@ -12,36 +12,46 @@ if (!is_dir('temp')) {
         die('Failed to create temp directory');
     }
 } else exec('rm -rf temp/*');
-$command = "unzip '$filename' 'assets/*/lang/en_us.json' -d temp";
-exec($command, $output, $return_var);
-$command = "unzip '$filename' 'assets/*/lang/hu_hu.json' -d temp";
-exec($command, $output, $return_var);
 
-// first directory is the key
-$dir = new DirectoryIterator('temp/assets');
-$keys = [];
-foreach ($dir as $fileinfo) {
-    if ($fileinfo->isDir() && !$fileinfo->isDot()) {
-        $keys[] = $fileinfo->getFilename();
-    }
+$files = ['en_us.json', 'hu_hu.json'];
+
+foreach($files as $file) {
+  $command = "unzip '$filename' 'assets/*/lang/'" . $file . " -d temp 2>/dev/null";
+  exec($command, $output, $return_var);
 }
 
-foreach($keys as $key) {
-  if (!is_dir("../originals/$key")) {
-      if (!mkdir("../originals/$key", 0755, true)) {
-          die('Failed to create directory');
+if (!is_dir('temp/assets')) {
+  echo "\e[033m ! No assets extracted from: " . $filename . "\e[0m" . PHP_EOL;
+  exit(0);
+}
+
+try {
+  $dir = new DirectoryIterator('temp/assets');
+  $keys = [];
+  foreach ($dir as $fileinfo) {
+      if ($fileinfo->isDir() && !$fileinfo->isDot()) {
+          $keys[] = $fileinfo->getFilename();
       }
   }
 
-  $files = ['en_us.json', 'hu_hu.json'];
-  foreach($files as $file)
-  {
-    $source = "temp/assets/$key/lang/$file";
-    $destination = "../originals/$key/$file";
-    if (!rename($source, $destination)) {
-        echo('Failed to move file: ' . $file . PHP_EOL);
+  foreach($keys as $key) {
+    if (!is_dir("../originals/$key")) {
+        if (!mkdir("../originals/$key", 0755, true)) {
+            die('Failed to create directory');
+        }
+    }
+
+    foreach($files as $file)
+    {
+      $source = "temp/assets/$key/lang/$file";
+      $destination = "../originals/$key/$file";
+      if (@rename($source, $destination)) echo "\e[032m   * Extracted: \e[034m$file\e[0m" . PHP_EOL;
     }
   }
+
+  exec('rm -rf temp/*');
+
+} catch (Exception $e) {
+  echo "\e[031m ! Error: " . $e->getMessage() . "\e[0m" . PHP_EOL;
 }
 
-exec('rm -rf temp/*');
